Fixed issue and refactored DefaultApiConfig
