+++ src/api/endpoints/chatMessage/ChatMessageEndpoint.ts
@@ -2,11 +2,11 @@ import { Logger } from "ts-log";
 
 import { HeaderProvider } from "../../../Func";
 import ChatMessage from "../../../models/ChatMessage";
-import { DefaultApiConfig } from "../../ApiConfig";
 import ApiResponse from "../../ApiResponse";
 import PaginatedResponse from "../../PaginatedResponse";
 import ApiEndpoint from "../ApiEndpoint";
 import TokenManager from "../auth/TokenManager";
+import DefaultApiConfig from "../DefaultApiConfig";
 import ChatMessageRequest from "./request/ChatMessageRequest";
 
 /**
