+++ src/api/endpoints/code/CodeEndpoint.ts
@@ -1,10 +1,10 @@
 import { Logger } from "ts-log";
 
 import { HeaderProvider } from "../../../Func";
-import { DefaultApiConfig } from "../../ApiConfig";
 import ApiResponse from "../../ApiResponse";
 import ApiEndpoint from "../ApiEndpoint";
 import TokenManager from "../auth/TokenManager";
+import DefaultApiConfig from "../DefaultApiConfig";
 
 /**
  * Endpoints for interacting with promotional codes.
