+++ src/api/endpoints/emergency/EmergencyEndpoint.ts
@@ -4,10 +4,10 @@ import { HeaderProvider } from "../../../Func";
 import { CancellationReason } from "../../../models/CancellationReason";
 import Emergency from "../../../models/Emergency";
 import { ResponseRating } from "../../../models/ResponseRating";
-import { DefaultApiConfig } from "../../ApiConfig";
 import ApiResponse from "../../ApiResponse";
 import ApiEndpoint from "../ApiEndpoint";
 import TokenManager from "../auth/TokenManager";
+import DefaultApiConfig from "../DefaultApiConfig";
 import CreateEmergencyRequest from "./request/CreateEmergencyRequest";
 import LocationDetail from "./response/LocationDetail";
 import TeamDetailsResponse from "./response/TeamDetailsResponse";
