+++ src/api/endpoints/staff/StaffEndpoint.ts
@@ -1,10 +1,10 @@
 import { Logger } from "ts-log";
 
 import { HeaderProvider } from "../../../Func";
-import { DefaultApiConfig } from "../../ApiConfig";
 import ApiResponse from "../../ApiResponse";
 import ApiEndpoint from "../ApiEndpoint";
 import TokenManager from "../auth/TokenManager";
+import DefaultApiConfig from "../DefaultApiConfig";
 import MedalInformation from "./response/MedalInformation";
 
 /**
